Put search classes into search.py
