+++ lutris/gui/lutriswindow.py
@@ -37,7 +37,8 @@ from lutris.gui.widgets.gi_composites import GtkTemplate
 from lutris.gui.widgets.sidebar import LutrisSidebar
 from lutris.gui.widgets.utils import load_icon_theme, open_uri
 from lutris.runtime import ComponentUpdater, RuntimeUpdater
-from lutris.services.base import SERVICE_GAMES_LOADED, SERVICE_LOGIN, SERVICE_LOGOUT
+from lutris.services.base import BaseService, SERVICE_GAMES_LOADED, SERVICE_LOGIN, SERVICE_LOGOUT
+from lutris.search import GameSearch
 from lutris.services.lutris import LutrisService
 from lutris.util import datapath
 from lutris.util.jobs import COMPLETED_IDLE_TASK, AsyncCall, schedule_at_idle
